+++ connectivity/__init__.py
@@ -1,11 +1,13 @@
 from .core import connectedness, beri
-from .rastio import overview_info, read_raster
+from .rastio import overview_info, pixel_coverage, read_raster, resolution_info
 from .utils import remove_grid_bias
 
 __all__ = [
     "connectedness",
     "beri",
     "overview_info",
+    "pixel_coverage",
     "read_raster",
     "remove_grid_bias",
+    "resolution_info",
 ]